fix: resolve todos 026/027 — agent-native parity, file splitting
Agent-native parity (Todo 026):
- operating-md.template.ts: merged workspace template content — session start
  protocol, channel lead duties (TeamCreate/TaskCreate/SendMessage), coordination
  messages, shift handoffs, memory protocol
- scaffold-payload.service.ts: added RULES.md, good-outputs.md, knowledge-index.yaml
  to agent scaffold output

File splitting (Todo 027):
- Split runner-event-router.ts from 868 → 445 lines by extracting 4 modules:
  - cli-stream-router.ts (172 lines): CLI stream-json event handling
  - sse-forward-handler.ts (128 lines): SSE forwarding + preview URLs
  - file-change-handler.ts (65 lines): Neon sync + file_changed broadcast
  - coordination-router.ts (59 lines): cross-channel coordination + @mentions
